split into table meta mod
