+++ vm/src/table_meta/mod.rs
@@ -1,13 +1,19 @@
+mod effects;
+mod find_address_visitor;
+use effects::WriteCache;
+use find_address_visitor::FindingAddressVisitor;
+
 use std::borrow::Borrow;
 use std::fmt::Formatter;
 use std::str::FromStr;
-use std::{collections::BTreeMap, fmt, fmt::Display};
+use std::{collections::BTreeMap, fmt::Display};
 
 use crate::natives::table::{TableChangeSet, TableHandle};
 use crate::session::SessionExt;
 use crate::size_change_set::{AccountSizeChangeSet, SizeChangeSet, SizeDelta};
 use crate::storage::data_view_resolver::TableMetaResolver;
 use anyhow::bail;
+
 use move_deps::move_core_types::effects::ChangeSet;
 use move_deps::move_core_types::language_storage::TypeTag;
 use move_deps::move_core_types::resolver::MoveResolver;
@@ -15,12 +21,11 @@ use move_deps::move_core_types::value::MoveTypeLayout;
 use move_deps::{
     move_binary_format::errors::{Location, PartialVMError, PartialVMResult, VMResult},
     move_core_types::{account_address::AccountAddress, effects::Op, vm_status::StatusCode},
-    move_vm_types::{
-        values::Value,
-        views::{ValueView, ValueVisitor},
-    },
+    move_vm_types::{values::Value, views::ValueView},
 };
 use serde::{Deserialize, Serialize};
+
+use self::effects::{deserialize_owner, deserialize_size, serialize_op, WriteCacheValue};
 
 #[derive(Clone, Eq, PartialEq, Hash, Ord, PartialOrd, Debug, Serialize, Deserialize)]
 pub enum TableMetaType {
@@ -82,7 +87,7 @@ pub fn resolve_table_ownership<S: TableMetaResolver + MoveResolver>(
     session: SessionExt<'_, '_, S>,
     change_set: &ChangeSet,
     table_change_set: &TableChangeSet,
-    table_owner_data_cache: &mut TableOwnerDataCache<'_, S>,
+    data_cache: &mut TableMetaDataCache<'_, S>,
 ) -> VMResult<()> {
     let new_tables = &table_change_set.borrow().new_tables;
 
@@ -90,8 +95,8 @@ pub fn resolve_table_ownership<S: TableMetaResolver + MoveResolver>(
     println!("remove table {:?}", table_change_set.removed_tables);
 
     for i in table_change_set.removed_tables.iter() {
-        table_owner_data_cache.del_owner(i);
-        table_owner_data_cache.del_size(i);
+        data_cache.del_owner(i);
+        data_cache.del_size(i);
     }
 
     for (addr, account_change_set) in change_set.borrow().accounts().iter() {
@@ -107,9 +112,9 @@ pub fn resolve_table_ownership<S: TableMetaResolver + MoveResolver>(
                 // already stored table's handle
 
                 if new_tables.contains_key(&found_handle)
-                    || table_owner_data_cache.is_registerd_table(&found_handle)?
+                    || data_cache.is_registerd_table(&found_handle)?
                 {
-                    table_owner_data_cache.set_owner(&found_handle, *addr);
+                    data_cache.set_owner(&found_handle, *addr);
                 }
             }
         }
@@ -122,9 +127,9 @@ pub fn resolve_table_ownership<S: TableMetaResolver + MoveResolver>(
             for address_found in res.iter() {
                 let found_handle = TableHandle(*address_found);
                 if new_tables.contains_key(&found_handle)
-                    || table_owner_data_cache.is_registerd_table(&found_handle)?
+                    || data_cache.is_registerd_table(&found_handle)?
                 {
-                    table_owner_data_cache.set_owner(&found_handle, outer_handle.0);
+                    data_cache.set_owner(&found_handle, outer_handle.0);
                 }
             }
         }
@@ -135,7 +140,7 @@ pub fn resolve_table_ownership<S: TableMetaResolver + MoveResolver>(
 pub fn resolve_table_size_change_by_account<S: TableMetaResolver>(
     table_change_set: &TableChangeSet,
     table_size_change: &SizeChangeSet<TableHandle>,
-    table_owner_data_cache: &mut TableOwnerDataCache<'_, S>,
+    data_cache: &mut TableMetaDataCache<'_, S>,
 ) -> VMResult<AccountSizeChangeSet> {
     let mut accounts_delta: AccountSizeChangeSet = AccountSizeChangeSet::default();
 
@@ -143,8 +148,8 @@ pub fn resolve_table_size_change_by_account<S: TableMetaResolver>(
     println!("remove table {:?}", table_change_set.removed_tables);
 
     // handle previous size
-    for (handle, owner) in table_owner_data_cache.get_owner_changes().into_iter() {
-        let size = match table_owner_data_cache.get_size(handle)? {
+    for (handle, owner) in data_cache.get_owner_changes().into_iter() {
+        let size = match data_cache.get_size(handle)? {
             Some(s) => s,
             None => {
                 continue;
@@ -153,11 +158,11 @@ pub fn resolve_table_size_change_by_account<S: TableMetaResolver>(
 
         match owner {
             Some(new_owner) => {
-                let old_owner = table_owner_data_cache.get_root_owner(handle)?.unwrap();
+                let old_owner = data_cache.get_root_owner(handle)?.unwrap();
                 accounts_delta.move_size(old_owner, new_owner.clone(), size);
             }
             None => {
-                let acc = table_owner_data_cache.get_root_owner(handle)?.unwrap();
+                let acc = data_cache.get_root_owner(handle)?.unwrap();
                 let delta = SizeDelta::decreasing(size);
                 accounts_delta.insert_size(acc, delta);
             }
@@ -171,11 +176,11 @@ pub fn resolve_table_size_change_by_account<S: TableMetaResolver>(
         if table_change_set.removed_tables.contains(handle) {
             continue;
         }
-        let owner = table_owner_data_cache.get_root_owner(handle)?.unwrap();
+        let owner = data_cache.get_root_owner(handle)?.unwrap();
 
         accounts_delta.insert_size(owner, size_delta.clone());
 
-        let current_size = table_owner_data_cache.get_size(&handle)?.unwrap_or(0);
+        let current_size = data_cache.get_size(&handle)?.unwrap_or(0);
 
         let mut delta = SizeDelta::increasing(current_size);
         delta.merge(size_delta.clone());
@@ -184,79 +189,12 @@ pub fn resolve_table_size_change_by_account<S: TableMetaResolver>(
         }
         let new_size = delta.amount;
 
-        table_owner_data_cache.set_size(&handle, new_size);
+        data_cache.set_size(&handle, new_size);
     }
 
     println!("{:?}", accounts_delta);
 
     Ok(accounts_delta)
-}
-
-struct FindingAddressVisitor {
-    addresses: Vec<AccountAddress>,
-}
-
-impl<'a> FindingAddressVisitor {
-    fn new() -> Self {
-        Self {
-            addresses: Vec::default(),
-        }
-    }
-
-    fn finish(self) -> Vec<AccountAddress> {
-        self.addresses
-    }
-}
-
-impl<'a> ValueVisitor for FindingAddressVisitor {
-    #[inline]
-    fn visit_u8(&mut self, _depth: usize, _val: u8) {}
-
-    #[inline]
-    fn visit_u64(&mut self, _depth: usize, _val: u64) {}
-
-    #[inline]
-    fn visit_u128(&mut self, _depth: usize, _val: u128) {}
-
-    #[inline]
-    fn visit_bool(&mut self, _depth: usize, _val: bool) {}
-
-    #[inline]
-    fn visit_address(&mut self, _depth: usize, _val: AccountAddress) {
-        self.addresses.push(_val);
-    }
-
-    #[inline]
-    fn visit_struct(&mut self, _depth: usize, _len: usize) -> bool {
-        true
-    }
-
-    #[inline]
-    fn visit_vec(&mut self, _depth: usize, _len: usize) -> bool {
-        true
-    }
-
-    #[inline]
-    fn visit_vec_u8(&mut self, _depth: usize, _vals: &[u8]) {}
-
-    #[inline]
-    fn visit_vec_u64(&mut self, _depth: usize, _vals: &[u64]) {}
-
-    #[inline]
-    fn visit_vec_u128(&mut self, _depth: usize, _vals: &[u128]) {}
-
-    #[inline]
-    fn visit_vec_bool(&mut self, _depth: usize, _vals: &[bool]) {}
-
-    #[inline]
-    fn visit_vec_address(&mut self, _depth: usize, vals: &[AccountAddress]) {
-        self.addresses.append(vals.to_vec().as_mut());
-    }
-
-    #[inline]
-    fn visit_ref(&mut self, _depth: usize, _is_global: bool) -> bool {
-        false
-    }
 }
 
 #[derive(Default)]
@@ -265,13 +203,13 @@ pub struct TableMetaChangeSet {
     pub size: BTreeMap<TableHandle, Op<Vec<u8>>>,
 }
 
-pub struct TableOwnerDataCache<'r, S> {
+pub struct TableMetaDataCache<'r, S> {
     remote: &'r S,
     table_owner: BTreeMap<TableHandle, WriteCache>,
     table_size: BTreeMap<TableHandle, WriteCache>,
 }
 
-impl<'r, S: TableMetaResolver> TableOwnerDataCache<'r, S> {
+impl<'r, S: TableMetaResolver> TableMetaDataCache<'r, S> {
     pub fn new(remote: &'r S) -> Self {
         Self {
             remote,
@@ -382,107 +320,4 @@ impl<'r, S: TableMetaResolver> TableOwnerDataCache<'r, S> {
 
         Ok(child_or_me)
     }
-}
-
-#[derive(Debug)]
-enum WriteCache {
-    /// A resource resides in this slot and will be write into storage.
-    Updated(WriteCacheValue),
-    /// A resource used to exist in storage but has been deleted by the current transaction.
-    Deleted,
-}
-
-impl WriteCache {
-    fn get_size(&self) -> Option<&usize> {
-        match self {
-            WriteCache::Updated(val) => match val {
-                WriteCacheValue::Owner(_) => panic!("not found"),
-                WriteCacheValue::Size(size) => Some(size),
-            },
-            WriteCache::Deleted => None,
-        }
-    }
-
-    fn get_owner(&self) -> Option<&AccountAddress> {
-        match self {
-            WriteCache::Updated(val) => match val {
-                WriteCacheValue::Owner(owner) => Some(owner),
-                WriteCacheValue::Size(_) => panic!("not"),
-            },
-            WriteCache::Deleted => None,
-        }
-    }
-
-    fn into_effect(self) -> Option<Op<WriteCacheValue>> {
-        match self {
-            Self::Deleted => Some(Op::Delete),
-            Self::Updated(val) => Some(Op::Modify(val)),
-        }
-    }
-}
-
-impl Display for WriteCache {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            WriteCache::Updated(val) => write!(f, "Updated({})", val),
-            WriteCache::Deleted => write!(f, "Deleted"),
-        }
-    }
-}
-
-#[derive(Debug)]
-enum WriteCacheValue {
-    Owner(AccountAddress),
-    Size(usize),
-}
-
-impl WriteCacheValue {
-    fn serialize(self) -> Option<Vec<u8>> {
-        match self {
-            WriteCacheValue::Owner(owner) => serialize_owner(&owner),
-            WriteCacheValue::Size(size) => serialize_size(&size),
-        }
-    }
-}
-
-impl Display for WriteCacheValue {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            WriteCacheValue::Owner(val) => write!(f, "Owner({})", val),
-            WriteCacheValue::Size(val) => write!(f, "Size({})", val),
-        }
-    }
-}
-
-fn serialize_op(op: Op<WriteCacheValue>) -> PartialVMResult<Op<Vec<u8>>> {
-    match op {
-        Op::New(v) => Ok(Op::New(serialize_or_error(v)?)),
-        Op::Modify(v) => Ok(Op::Modify(serialize_or_error(v)?)),
-        Op::Delete => Ok(Op::Delete),
-    }
-}
-
-fn serialize_or_error(v: WriteCacheValue) -> PartialVMResult<Vec<u8>> {
-    match v.serialize() {
-        Some(d) => Ok(d),
-        None => Err(PartialVMError::new(
-            StatusCode::FAILED_TO_SERIALIZE_WRITE_SET_CHANGES,
-        )),
-    }
-}
-
-fn serialize_owner(owner: &AccountAddress) -> Option<Vec<u8>> {
-    bcs::to_bytes(owner).ok()
-}
-
-fn deserialize_owner(blob: &[u8]) -> Option<AccountAddress> {
-    bcs::from_bytes(blob).ok()
-}
-
-fn serialize_size(size: &usize) -> Option<Vec<u8>> {
-    bcs::to_bytes(size).ok()
-}
-
-fn deserialize_size(blob: &[u8]) -> Option<usize> {
-    bcs::from_bytes(blob).ok()
 }
